+++ CHANGELOG.md
@@ -2,6 +2,62 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="3.199.0"></a>
+# [3.199.0](https://github.com/PeerioTechnologies/peerio-mobile/compare/v3.198.0...v3.199.0) (2017-11-10)
+
+
+### Bug Fixes
+
+* image size preference should be in reverse ([fef6ce6](https://github.com/PeerioTechnologies/peerio-mobile/commit/fef6ce6))
+
+
+
+<a name="3.198.0"></a>
+# [3.198.0](https://github.com/PeerioTechnologies/peerio-mobile/compare/v3.197.0...v3.198.0) (2017-11-10)
+
+
+### Features
+
+* always expand content when content is enabled ([40dafc6](https://github.com/PeerioTechnologies/peerio-mobile/commit/40dafc6))
+
+
+
+<a name="3.197.0"></a>
+# [3.197.0](https://github.com/PeerioTechnologies/peerio-mobile/compare/v3.190.0...v3.197.0) (2017-11-10)
+
+
+### Bug Fixes
+
+* display image limits ([7d0a37f](https://github.com/PeerioTechnologies/peerio-mobile/commit/7d0a37f))
+* navigate folders in file share ([95c342a](https://github.com/PeerioTechnologies/peerio-mobile/commit/95c342a))
+* pr template [skip ci] ([150a1a7](https://github.com/PeerioTechnologies/peerio-mobile/commit/150a1a7))
+* restore new line at the end of file ([8f3224f](https://github.com/PeerioTechnologies/peerio-mobile/commit/8f3224f))
+
+
+### Features
+
+* dependecy check ([640b339](https://github.com/PeerioTechnologies/peerio-mobile/commit/640b339))
+
+
+
+<a name="3.191.0"></a>
+# [3.191.0](https://github.com/PeerioTechnologies/peerio-mobile/compare/v3.190.0...v3.191.0) (2017-11-10)
+
+
+### Bug Fixes
+
+* display image limits ([7d0a37f](https://github.com/PeerioTechnologies/peerio-mobile/commit/7d0a37f))
+* navigate folders in file share ([95c342a](https://github.com/PeerioTechnologies/peerio-mobile/commit/95c342a))
+* pr template [skip ci] ([150a1a7](https://github.com/PeerioTechnologies/peerio-mobile/commit/150a1a7))
+* restore new line at the end of file ([8f3224f](https://github.com/PeerioTechnologies/peerio-mobile/commit/8f3224f))
+
+
+### Features
+
+* dependecy check ([640b339](https://github.com/PeerioTechnologies/peerio-mobile/commit/640b339))
+
+
+
 <a name="3.190.0"></a>
 # [3.190.0](https://github.com/PeerioTechnologies/peerio-mobile/compare/v3.189.0...v3.190.0) (2017-11-08)